chore: refactor structure styles
