+++ src/components/elements/index.js
@@ -1,5 +1,5 @@
 export * from './button'
 export * from './container'
 export * from './image'
-export * from './input'
 export * from './typography'
+export * from './input'
